問題出たunsignedをNSIntegerに変更、いずれほとんど置き換える予定

+++ Thousand/Thousand App/Standard Plugins/THSecondaryThreadProcessor.h
@@ -48,7 +48,7 @@
 -(NSArray *)uniqueNamesOfdependingPlugins ;
 
 	//protocol T2ThreadProcessing_v100
--(void)processThread:(T2Thread *)thread appendingIndex:(unsigned)index ;
+-(void)processThread:(T2Thread *)thread appendingIndex:(NSInteger)index ;
 
 	//protocol T2ResExtracting_v100
 -(NSArray *)extractKeys ;
